Strategy design patter
