+++ lowleveldesign-programming/src/main/java/Decorate/BasicCar.java
@@ -1,0 +1,9 @@
+package Decorate;
+
+public class BasicCar implements Car {
+
+    @Override
+    public void assemble() {
+        System.out.println("Basic Car.");
+    }
+}
