+++ lowleveldesign-programming/src/main/java/Decorate/Car.java
@@ -1,0 +1,6 @@
+package Decorate;
+
+public interface Car {
+
+    public void assemble();
+}
