+++ lowleveldesign-programming/src/main/java/Decorate/CarDecorator.java
@@ -1,0 +1,14 @@
+package Decorate;
+
+public class CarDecorator implements Car {
+
+    Car car;
+    public CarDecorator(Car car){
+       this.car = car;
+    }
+    @Override
+    public void assemble() {
+        this.car.assemble();
+        System.out.println(" Adding features of Sports Car.");
+    }
+}
