+++ lowleveldesign-programming/src/main/java/Decorate/LuxuryCar.java
@@ -1,0 +1,12 @@
+package Decorate;
+
+public class LuxuryCar extends CarDecorator {
+    public LuxuryCar(Car car) {
+        super(car);
+    }
+
+    @Override
+    public void assemble() {
+        super.assemble();
+    }
+}
