+++ lowleveldesign-programming/src/main/java/Decorate/SportsCar.java
@@ -1,0 +1,16 @@
+package Decorate;
+
+public class SportsCar extends CarDecorator {
+
+    public SportsCar(Car car){
+
+        super(car);
+
+    }
+
+    @Override
+    public void assemble() {
+        super.assemble();
+        System.out.println(" Adding features of Sports Car.");
+    }
+}
