+++ lowleveldesign-programming/src/main/java/Decorate/Test.java
@@ -1,0 +1,13 @@
+package Decorate;
+
+public class Test {
+
+    public static void main(String[] args) {
+        Car sportsCar = new SportsCar(new BasicCar());
+        sportsCar.assemble();
+
+        Car sportsLuxuryCar = new SportsCar(new LuxuryCar(new BasicCar()));
+        sportsLuxuryCar.assemble();
+
+    }
+}
