--- lowleveldesign-programming/src/main/java/com/app/design/patterns/AbstractFactory/Button.java
@@ -1,6 +1,0 @@
-package com.app.design.patterns.AbstractFactory;
-
-public interface Button {
-
-    public void render();
-}
