--- lowleveldesign-programming/src/main/java/com/app/design/patterns/AbstractFactory/CheckBox.java
@@ -1,6 +1,0 @@
-package com.app.design.patterns.AbstractFactory;
-
-public interface CheckBox {
-
-    void render();
-}
