--- lowleveldesign-programming/src/main/java/com/app/design/patterns/AbstractFactory/GuiFactory.java
@@ -1,7 +1,0 @@
-package com.app.design.patterns.AbstractFactory;
-
-public interface GuiFactory {
-
-    Button crteateButton();
-    CheckBox createCheckBox();
-}
