--- lowleveldesign-programming/src/main/java/com/app/design/patterns/AbstractFactory/MacBookButton.java
@@ -1,8 +1,0 @@
-package com.app.design.patterns.AbstractFactory;
-
-public class MacBookButton implements Button{
-    @Override
-    public void render() {
-        System.out.println("Rendering a Macbook button");
-    }
-}
