--- lowleveldesign-programming/src/main/java/com/app/design/patterns/AbstractFactory/MacBookCheckBox.java
@@ -1,8 +1,0 @@
-package com.app.design.patterns.AbstractFactory;
-
-public class MacBookCheckBox implements CheckBox{
-    @Override
-    public void render() {
-        System.out.println("Rendering a macbook checkbox");
-    }
-}
