--- lowleveldesign-programming/src/main/java/com/app/design/patterns/AbstractFactory/MacbookFactory.java
@@ -1,13 +1,0 @@
-package com.app.design.patterns.AbstractFactory;
-
-public class MacbookFactory implements GuiFactory{
-    @Override
-    public Button crteateButton() {
-        return new MacBookButton();
-    }
-
-    @Override
-    public CheckBox createCheckBox() {
-        return new MacBookCheckBox();
-    }
-}
