--- lowleveldesign-programming/src/main/java/com/app/design/patterns/AbstractFactory/WindowButton.java
@@ -1,8 +1,0 @@
-package com.app.design.patterns.AbstractFactory;
-
-public class WindowButton implements Button {
-    @Override
-    public void render() {
-        System.out.println("Rendering a Windows button");
-    }
-}
