--- lowleveldesign-programming/src/main/java/com/app/design/patterns/AbstractFactory/WindowFactory.java
@@ -1,13 +1,0 @@
-package com.app.design.patterns.AbstractFactory;
-
-public class WindowFactory implements GuiFactory{
-    @Override
-    public Button crteateButton() {
-        return new WindowButton();
-    }
-
-    @Override
-    public CheckBox createCheckBox() {
-        return new WindowCheckBox();
-    }
-}
