+++ lowleveldesign-programming/src/main/java/com/app/design/patterns/CreationalDesignPatter/AbstractFactory/Button.java
@@ -1,0 +1,6 @@
+package com.app.design.patterns.CreationalDesignPatter.AbstractFactory;
+
+public interface Button {
+
+    public void render();
+}
