+++ lowleveldesign-programming/src/main/java/com/app/design/patterns/CreationalDesignPatter/AbstractFactory/GuiFactory.java
@@ -1,0 +1,7 @@
+package com.app.design.patterns.CreationalDesignPatter.AbstractFactory;
+
+public interface GuiFactory {
+
+    Button crteateButton();
+    CheckBox createCheckBox();
+}
