+++ lowleveldesign-programming/src/main/java/com/app/design/patterns/CreationalDesignPatter/AbstractFactory/WindowButton.java
@@ -1,0 +1,8 @@
+package com.app.design.patterns.CreationalDesignPatter.AbstractFactory;
+
+public class WindowButton implements Button {
+    @Override
+    public void render() {
+        System.out.println("Rendering a Windows button");
+    }
+}
