+++ lowleveldesign-programming/src/main/java/com/app/design/patterns/CreationalDesignPatter/AbstractFactory/WindowCheckBox.java
@@ -1,0 +1,8 @@
+package com.app.design.patterns.CreationalDesignPatter.AbstractFactory;
+
+public class WindowCheckBox implements CheckBox {
+    @Override
+    public void render() {
+        System.out.println("Rendering a Windows checkbox");
+    }
+}
