+++ lowleveldesign-programming/src/main/java/com/app/design/patterns/CreationalDesignPatter/Factory/EmailNotification.java
@@ -1,0 +1,9 @@
+package com.app.design.patterns.CreationalDesignPatter.Factory;
+
+public class EmailNotification implements Notification {
+    @Override
+    public void notifyUser() {
+
+        System.out.println("Sending Email notification");
+    }
+}
