+++ lowleveldesign-programming/src/main/java/com/app/design/patterns/CreationalDesignPatter/Factory/Notification.java
@@ -1,0 +1,8 @@
+package com.app.design.patterns.CreationalDesignPatter.Factory;
+
+public interface Notification {
+
+    public void notifyUser();
+
+
+}
