+++ lowleveldesign-programming/src/main/java/com/app/design/patterns/CreationalDesignPatter/Factory/PushNotification.java
@@ -1,0 +1,8 @@
+package com.app.design.patterns.CreationalDesignPatter.Factory;
+
+public class PushNotification implements Notification {
+    @Override
+    public void notifyUser() {
+        System.out.println("Sending an Push notification");
+    }
+}
