+++ lowleveldesign-programming/src/main/java/com/app/design/patterns/CreationalDesignPatter/Factory/SMSNotification.java
@@ -1,0 +1,9 @@
+package com.app.design.patterns.CreationalDesignPatter.Factory;
+
+public class SMSNotification implements Notification {
+    @Override
+    public void notifyUser() {
+
+        System.out.println("Sending an SMS notification");
+    }
+}
