--- lowleveldesign-programming/src/main/java/com/app/design/patterns/Decorate/BasicCar.java
@@ -1,9 +1,0 @@
-package com.app.design.patterns.Decorate;
-
-public class BasicCar implements Car {
-
-    @Override
-    public void assemble() {
-        System.out.println("Basic Car.");
-    }
-}
