--- lowleveldesign-programming/src/main/java/com/app/design/patterns/Decorate/Car.java
@@ -1,6 +1,0 @@
-package com.app.design.patterns.Decorate;
-
-public interface Car {
-
-    public void assemble();
-}
