--- lowleveldesign-programming/src/main/java/com/app/design/patterns/Decorate/LuxuryCar.java
@@ -1,12 +1,0 @@
-package com.app.design.patterns.Decorate;
-
-public class LuxuryCar extends CarDecorator{
-    public LuxuryCar(Car car) {
-        super(car);
-    }
-
-    @Override
-    public void assemble() {
-        super.assemble();
-    }
-}
