--- lowleveldesign-programming/src/main/java/com/app/design/patterns/Decorate/SportsCar.java
@@ -1,16 +1,0 @@
-package com.app.design.patterns.Decorate;
-
-public class SportsCar extends CarDecorator {
-
-    public SportsCar(Car car){
-
-        super(car);
-
-    }
-
-    @Override
-    public void assemble() {
-        super.assemble();
-        System.out.println(" Adding features of Sports Car.");
-    }
-}
