--- lowleveldesign-programming/src/main/java/com/app/design/patterns/Decorate/Test.java
@@ -1,13 +1,0 @@
-package com.app.design.patterns.Decorate;
-
-public class Test {
-
-    public static void main(String[] args) {
-        Car sportsCar = new SportsCar(new BasicCar());
-        sportsCar.assemble();
-
-        Car sportsLuxuryCar = new SportsCar(new LuxuryCar(new BasicCar()));
-        sportsLuxuryCar.assemble();
-
-    }
-}
