--- lowleveldesign-programming/src/main/java/com/app/design/patterns/Factory/EmailNotification.java
@@ -1,9 +1,0 @@
-package com.app.design.patterns.Factory;
-
-public class EmailNotification implements Notification{
-    @Override
-    public void notifyUser() {
-
-        System.out.println("Sending Email notification");
-    }
-}
