--- lowleveldesign-programming/src/main/java/com/app/design/patterns/Factory/Notification.java
@@ -1,8 +1,0 @@
-package com.app.design.patterns.Factory;
-
-public interface Notification {
-
-    public void notifyUser();
-
-
-}
