--- lowleveldesign-programming/src/main/java/com/app/design/patterns/Factory/PushNotification.java
@@ -1,8 +1,0 @@
-package com.app.design.patterns.Factory;
-
-public class PushNotification implements Notification {
-    @Override
-    public void notifyUser() {
-        System.out.println("Sending an Push notification");
-    }
-}
