--- lowleveldesign-programming/src/main/java/com/app/design/patterns/Factory/SMSNotification.java
@@ -1,9 +1,0 @@
-package com.app.design.patterns.Factory;
-
-public class SMSNotification implements Notification{
-    @Override
-    public void notifyUser() {
-
-        System.out.println("Sending an SMS notification");
-    }
-}
