+++ lowleveldesign-programming/src/main/java/strategy/BinarySearch.java
@@ -1,0 +1,12 @@
+package strategy;
+
+import java.util.List;
+
+public class BinarySearch implements Searcher {
+    @Override
+    public Integer search(int key, List<Integer> nums) {
+        System.out.println(" From Binary-Searcher ");
+
+        return 1;
+    }
+}
