+++ lowleveldesign-programming/src/main/java/strategy/Inserter.java
@@ -1,0 +1,7 @@
+package strategy;
+
+import java.util.List;
+
+public interface Inserter {
+    public void insert(int num, List<Integer> nums);
+}
