+++ lowleveldesign-programming/src/main/java/strategy/LinearSearch.java
@@ -1,0 +1,12 @@
+package strategy;
+
+import java.util.List;
+
+public class LinearSearch implements Searcher {
+    @Override
+    public Integer search(int key, List<Integer> nums) {
+        System.out.println(" From Linear-Searcher ");
+
+        return 1;
+    }
+}
