+++ lowleveldesign-programming/src/main/java/strategy/Searcher.java
@@ -1,0 +1,7 @@
+package strategy;
+
+import java.util.List;
+
+public interface Searcher {
+    public Integer search(int key, List<Integer> nums);
+}
